Move the firewall factory to the correct namespace
